requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
